+++ Icon/Icon.js
@@ -338,6 +338,8 @@ const IconBase = kind({
  * bluetooth
  * moodmode
  * changepassword
+ * r2rappcall
+ * musicsrc
  * ```
  *
  * @name iconList