+++ Input/Input.js
@@ -372,7 +372,8 @@ const InputPopupBase = kind({
 		minLength,
 		...rest
 	}) => {
-
+		const id = `inputPopup`;
+		const ariaLabelledBy = popupAriaLabel ? null : `${id}_title ${id}_subtitle`;
 		const inputProps = extractInputFieldProps(rest);
 		const numberMode = (numberInputField !== 'field') && (type === 'number' || type === 'passwordnumber');
 		// Set up the back button
@@ -386,7 +387,7 @@ const InputPopupBase = kind({
 				size="small"
 			/>
 		) : null);
-		const heading = <Heading size="title" marqueeOn="render" alignment="center" className={css.title}>{title}</Heading>;
+		const heading = <Heading id={`${id}_title`} size="title" marqueeOn="render" alignment="center" className={css.title}>{title}</Heading>;
 
 		delete rest.length;
 		delete rest.onComplete;
@@ -395,12 +396,15 @@ const InputPopupBase = kind({
 		return (
 			<Popup
 				aria-label={popupAriaLabel}
+				aria-labelledby={ariaLabelledBy}
 				onClose={onClose}
 				onShow={onShow}
 				position={popupType === 'fullscreen' ? 'fullscreen' : 'center'}
 				className={popupClassName}
+				noAlertRole
 				noAnimation
 				open={open}
+				role="region"
 			>
 				{popupType === 'fullscreen' ? backButton : null}
 				<Layout orientation="vertical" align={`center ${numberMode ? 'space-between' : ''}`} className={css.body}>
@@ -412,7 +416,7 @@ const InputPopupBase = kind({
 								{heading}
 							</>
 						}
-						<Heading size="subtitle" marqueeOn="render" alignment="center" className={css.subtitle}>{subtitle}</Heading>
+						<Heading id={`${id}_subtitle`} size="subtitle" marqueeOn="render" alignment="center" className={css.subtitle}>{subtitle}</Heading>
 					</Cell>
 					<Cell shrink className={css.inputArea}>
 						{numberMode ?
@@ -439,7 +443,6 @@ const InputPopupBase = kind({
 								autoFocus
 								type={type}
 								defaultValue={value}
-								noReadoutOnFocus
 								placeholder={placeholder}
 								onBeforeChange={onBeforeChange}
 								onKeyDown={onInputKeyDown}